fixing tsc
