+++ package.json
@@ -17,7 +17,7 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.30",
     "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.11",
-    "@palmetto/palmetto-design-tokens": "^0.12.2",
+    "@palmetto/palmetto-design-tokens": "^0.13.0",
     "@types/cleave.js": "^1.4.3",
     "react-select": "^3.1.0"
   },
