+++ src/components/Table/Table.stories.mdx
@@ -1,7 +1,6 @@
 import { useState, useEffect } from 'react';
 import { action } from '@storybook/addon-actions';
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
-import { Column } from './TableTypes';
 import Button from '../Button/Button';
 import Table from './Table';
