+++ src/components/Table/TableRow/TableRow.tsx
@@ -1,7 +1,7 @@
 import React, { FC, ReactNode } from 'react';
 import classNames from 'classnames';
 import styles from './TableRow.module.scss';
-import { Column, EventWithColumnKey, Row } from '../TableTypes';
+import { Column, EventWithColumnKey, Row } from '../../../types';
 import getColumnKeys from '../../../lib/getColumnKeys';
 import TableCell from '../TableCell/TableCell';
 import TableHeaderCell from '../TableHeaderCell/TableHeaderCell';
